# Conflicts:
#	build.sh
#	config.xml

+++ www/js/class-invoice.js
@@ -113,9 +113,21 @@
 		this.recalc();
 	}
 	Invoice.prototype.addLine = function(item){
-		this.salesLines.push(item);
-		NativeStorage.setItem('cart', JSON.stringify(this.salesLines), nsSetNoop, noop);
-		this.recalc();
+		var inCartAlready = false;
+		for (var i = 0; i < this.salesLines.length; i++) {
+			if(this.salesLines[i].sku == item.sku && this.salesLines[i].location == item.location){
+				inCartAlready = true;
+				this.salesLines[i].quantity = parseInt(this.salesLines[i].quantity) + parseInt(item.quantity);
+				NativeStorage.setItem('cart', JSON.stringify(this.salesLines), nsSetNoop, noop);
+				this.recalc();
+				break;
+			}
+		}
+		if(!inCartAlready){
+			this.salesLines.push(item);
+			NativeStorage.setItem('cart', JSON.stringify(this.salesLines), nsSetNoop, noop);
+			this.recalc();
+		}
 	}
 	Invoice.prototype.deleteLine = function(id){
 		this.salesLines.splice(id, 1);
@@ -347,8 +359,7 @@
 
 	// TODO: Show Discounts
 	Invoice.prototype.draw = function(selector){
-		//if(this.salesLines.length == 0){
-		if(false){
+		if(this.salesLines.length == 0){
 			if (selector === undefined || selector == '.cart-list'){
 				var newSaleHTML = '<div id="new-sale"><div id="new-sale-header"></div><div id="returning-customer">Returning</div></div>';
 				$$('.cart-list').empty();
@@ -375,7 +386,7 @@
 				            clearTimeout(lookupDelayTimer);
 				            (function(search){
 				                lookupDelayTimer = setTimeout(function() {
-				                    TM.searchCustomer(search, function(data){
+				                    TM.searchAnyValue(search, function(data){
 				                        $$('.search-results-lookup').empty();
 				                        for (var i = 0; i < listNames.length; i++) {
 				                            //var searchItem = $$('<li class="item-content" data-sku="'+listSkus[i]+'"><div class="item-inner"><div class="item-title search-result-item">'+listNames[i]+'</div></div></li>');
@@ -385,7 +396,7 @@
 				                            });
 				                            //$$('.search-results-lookup').append(searchItem);
 				                        }
-				                    });
+									});
 				                }, 500); // Will do the ajax stuff after 1000 ms, or 1 s
 				            })(this.value);
 				        }
@@ -423,7 +434,7 @@
 							    		'<p class="center-align"><i class="icon f7-icons">bars</i></p>'+
 							    	'</div>' +
 							    	'<div class="col-33 quantity-col-card right-float right-align" data-id="'+i+'">' +
-							    		'<p>QTY:' + this.salesLines[i].quantity + '</p>'+
+							    		'<p><span class="qty-button-grey">' + this.salesLines[i].quantity + '<i class="fa fa-sort" aria-hidden="true"></i></span></p>'+
 							    	'</div>' +
 							    '</div>' +
 						    '</div>' +
